requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-crispy-forms==1.13.0
django-phone-field==1.8.1
idna==3.3
Pillow==9.0.0
python-dotenv==0.19.2
regex==2022.1.18
requests==2.27.1
sqlparse==0.4.2
stripe==2.65.0
tzdata==2021.5
urllib3==1.26.8
